ml-data-pipeline/requirements.txt
pandas>=2.2.0
pyarrow>=15.0.0
sqlalchemy>=2.0.0
pydantic>=2.5.0
aiohttp>=3.9.0
rich>=13.0.0
pyyaml>=6.0
python-dotenv>=1.0.0
pytest>=8.0.0
pytest-asyncio>=0.23.0
